fix the partially typed case fix icons, and Equals calls based on review update dependencies use IsDescendantOf, update "org.jetbrains.intellij"
