+++ rider/build.gradle.kts
@@ -11,11 +11,11 @@ repositories {
 }
 
 plugins {
-    id("org.jetbrains.intellij") version "1.7.0" // https://github.com/JetBrains/gradle-intellij-plugin/releases
+    id("org.jetbrains.intellij") version "1.8.0" // https://github.com/JetBrains/gradle-intellij-plugin/releases
     id("org.jetbrains.grammarkit") version "2021.2.2"
-    id("me.filippov.gradle.jvm.wrapper") version "0.10.0"
+    id("me.filippov.gradle.jvm.wrapper") version "0.11.0"
     id("com.jetbrains.rdgen") version "2022.2.0" // https://www.myget.org/feed/rd-snapshots/package/maven/com.jetbrains.rd/rd-gen
-    kotlin("jvm") version "1.6.10"
+    kotlin("jvm") version "1.7.0"
 }
 
 apply {
